+++ Library/Homebrew/keg.rb
@@ -72,13 +72,13 @@ class Keg
   # These paths relative to the keg's share directory should always be real
   # directories in the prefix, never symlinks.
   SHARE_PATHS = %w[
-    aclocal doc info locale man
+    aclocal doc info java locale man
     man/man1 man/man2 man/man3 man/man4
     man/man5 man/man6 man/man7 man/man8
     man/cat1 man/cat2 man/cat3 man/cat4
     man/cat5 man/cat6 man/cat7 man/cat8
     applications gnome gnome/help icons
-    mime-info pixmaps sounds
+    mime-info pixmaps sounds postgresql
   ]
 
   # if path is a file in a keg then this will return the containing Keg object
@@ -101,11 +101,6 @@ class Keg
     @name = path.parent.basename.to_s
     @linked_keg_record = HOMEBREW_LIBRARY.join("LinkedKegs", name)
     @opt_record = HOMEBREW_PREFIX.join("opt", name)
-  end
-
-  def fname
-    opoo "Keg#fname is a deprecated alias for Keg#name and will be removed soon"
-    name
   end
 
   def to_s
@@ -139,12 +134,32 @@ class Keg
     path.abv
   end
 
+  def disk_usage
+    path.disk_usage
+  end
+
+  def file_count
+    path.file_count
+  end
+
   def directory?
     path.directory?
   end
 
   def exist?
     path.exist?
+  end
+
+  def empty_installation?
+    Pathname.glob("#{path}/**/*") do |file|
+      next if file.directory?
+      basename = file.basename.to_s
+      next if Metafiles.copy?(basename)
+      next if %w[.DS_Store INSTALL_RECEIPT.json].include?(basename)
+      return false
+    end
+
+    true
   end
 
   def /(other)
@@ -183,9 +198,10 @@ class Keg
     path.rmtree
     path.parent.rmdir_if_possible
     remove_opt_record if optlinked?
+    remove_oldname_opt_record
   end
 
-  def unlink
+  def unlink(mode = OpenStruct.new)
     ObserverPathnameExtension.reset_counts!
 
     dirs = []
@@ -200,6 +216,12 @@ class Keg
 
         # check whether the file to be unlinked is from the current keg first
         if dst.symlink? && src == dst.resolved_path
+          if mode.dry_run
+            puts dst
+            Find.prune if src.directory?
+            next
+          end
+
           dst.uninstall_info if dst.to_s =~ INFOFILE_RX
           dst.unlink
           Find.prune if src.directory?
@@ -207,15 +229,22 @@ class Keg
       end
     end
 
-    remove_linked_keg_record if linked?
-
-    dirs.reverse_each(&:rmdir_if_possible)
+    unless mode.dry_run
+      remove_linked_keg_record if linked?
+      dirs.reverse_each(&:rmdir_if_possible)
+    end
 
     ObserverPathnameExtension.total
   end
 
   def lock
-    FormulaLock.new(name).with_lock { yield }
+    FormulaLock.new(name).with_lock do
+      if oldname_opt_record
+        FormulaLock.new(oldname_opt_record.basename.to_s).with_lock { yield }
+      else
+        yield
+      end
+    end
   end
 
   def completion_installed?(shell)
@@ -239,12 +268,18 @@ class Keg
     Dir["#{path}/lib/python2.7/site-packages/*.pth"].any?
   end
 
+  def apps
+    app_prefix = optlinked? ? opt_record : path
+    Pathname.glob("#{app_prefix}/{,libexec/}*.app")
+  end
+
   def app_installed?
-    Dir["#{path}/{,libexec/}*.app"].any?
+    !apps.empty?
   end
 
   def elisp_installed?
-    Dir["#{path}/share/emacs/site-lisp/**/*.el"].any?
+    return false unless (path/"share/emacs/site-lisp"/name).exist?
+    (path/"share/emacs/site-lisp"/name).children.any? { |f| %w[.el .elc].include? f.extname }
   end
 
   def version
@@ -254,6 +289,14 @@ class Keg
 
   def find(*args, &block)
     path.find(*args, &block)
+  end
+
+  def oldname_opt_record
+    @oldname_opt_record ||= if (opt_dir = HOMEBREW_PREFIX/"opt").directory?
+      opt_dir.subdirs.detect do |dir|
+        dir.symlink? && dir != opt_record && path.parent == dir.resolved_path.parent
+      end
+    end
   end
 
   def link(mode = OpenStruct.new)
@@ -279,6 +322,8 @@ class Keg
       when /^icons\// then :mkpath
       when /^zsh/ then :mkpath
       when /^fish/ then :mkpath
+      # Lua, Lua51, Lua53 all need the same handling.
+      when /^lua\// then :mkpath
       else :link
       end
     end
@@ -294,12 +339,15 @@ class Keg
       when "dtrace" then :mkpath
       when /^gdk-pixbuf/ then :mkpath
       when "ghc" then :mkpath
+      when /^gio/ then :mkpath
       when "lua" then :mkpath
+      when /^mecab/ then :mkpath
       when /^node/ then :mkpath
       when /^ocaml/ then :mkpath
       when /^perl5/ then :mkpath
       when "php" then :mkpath
       when /^python[23]\.\d/ then :mkpath
+      when /^R/ then :mkpath
       when /^ruby/ then :mkpath
       # Everything else is symlinked to the cellar
       else :link
@@ -329,9 +377,22 @@ class Keg
     ObserverPathnameExtension.total
   end
 
+  def remove_oldname_opt_record
+    return unless oldname_opt_record
+    return unless oldname_opt_record.resolved_path == path
+    @oldname_opt_record.unlink
+    @oldname_opt_record.parent.rmdir_if_possible
+    @oldname_opt_record = nil
+  end
+
   def optlink(mode = OpenStruct.new)
     opt_record.delete if opt_record.symlink? || opt_record.exist?
     make_relative_symlink(opt_record, path, mode)
+
+    if oldname_opt_record
+      oldname_opt_record.delete
+      make_relative_symlink(oldname_opt_record, path, mode)
+    end
   end
 
   def delete_pyc_files!
@@ -421,6 +482,7 @@ class Keg
 
       if src.symlink? || src.file?
         Find.prune if File.basename(src) == ".DS_Store"
+        Find.prune if src.realpath == dst
         # Don't link pyc files because Python overwrites these cached object
         # files and next time brew wants to link, the pyc file is in the way.
         if src.extname == ".pyc" && src.to_s =~ /site-packages/
